V2 of gui and pipeline
Contains additional features such as a console and loading bar on the gui, along with a message that pops up to tell you if the pipeline is completed. The pipeline changes contain more basic statements that show on the console to show the progress and the step that the pipeline is on to the user.